Simplify web/blob build
This merges web/Makefile and web/blob/Makefile, where the former's sole
purpose was to call the latter.

The the blob-stamp file can cause issues if it's present, but not the
blob/files.go file. As only the latter is the real dependency for the
build process, this indirection caused me some minutes to understand
why the build process was broken (don't ask me why I deleted files.go).
It can aslo be noted that all dependencies for blob/files.go are
outside of the blob/ directory, so the separation isn't even logically
necessary.

The initial Makefile setup has been added in 3e97a36.